Add files via upload
Principal Component Analysis is a technique that is used to transform the original dimensions of a dataset into a set of dimensions. 
One of the most useful aspects of PCA is to reduce dimensions. The idea behind this concept is that not all dimensions are equally important when describing the data, 
some of them explain the variation more than the others. Using this fact, we can identify which dimensions are more important to describe our data. 

In this example, we are using PCA to compress the image, while still maintaining it's quality to an acceptable level.