Add files via upload
This project was built as a final project for the course "Data Design and Representation".

We Scraped over 3000 recipe pages from the Food Network website to capture relevant information related to their diet, based on nutritional value, and find similar recipes based on categories and other celebrity chef recipes. We used Python's BeautifulSoup and Requests packages to scrape the website pages. Due to the huge volume of the data that we wanted to scrape we utilized several custom-built functions and built a scraping routine that would scrape as well as process the data to ensure that we have top-notch data quality, with standardized data formats, and also make it easier for users to query and access our data. We choose to use MongoDB over others to store the data due to its flexibility and high performance.  Overall, our web scraping and database design processes have resulted in a comprehensive recipe database that can be used to build a successful food recommender app.